"subtareasrobustas"

+++ src/subtasks/subtasks.controller.ts
@@ -1,4 +1,4 @@
-import { Controller,Get,Post,Param,Patch,Delete, Body } from '@nestjs/common';
+import { Controller,Get,Post,Param,Patch,Delete, Body,BadRequestException } from '@nestjs/common';
 import { SubtasksService } from './subtasks.service';
 import { CreateSubtaskDto } from './dto/CreateSubtask.dto';
 import { Auth } from 'src/auth/decorators/auth.decorator';
@@ -7,7 +7,6 @@ import { Role_User } from '@prisma/client';
 import { UsersService } from 'src/users/users.service';
 import { UserActiveInterface } from 'src/auth/interface/user-active.interface';
 import { ActiveUser } from 'src/users/decorators/active-user.decorator';
-import { get } from 'http';
 
 @Auth(Role_User.USER)
 @ApiTags("Subtasks")
@@ -22,6 +21,9 @@ export class SubtasksController {
 
     @Get(':id')
     async findOne(@Param('id') id: string){
+        if (!id) {
+            throw new BadRequestException('Subtask ID is required');
+        }
         return await this.subtaskService.findOne(id);
     }
 
@@ -31,12 +33,20 @@ export class SubtasksController {
     }
 
     @Patch(':id')
-    async update(@Param('id') id: string,@Body()createSubtaskDto:Partial<CreateSubtaskDto>,@ActiveUser()user:UserActiveInterface){
-        return await this.subtaskService.update(id,createSubtaskDto,user);
+    async update(
+        @Param('id') id: string,
+        @Body()createSubtaskDto:Partial<CreateSubtaskDto>,@ActiveUser()user:UserActiveInterface){
+            if (!id) {
+                throw new BadRequestException('Subtask ID is required');
+            }        
+            return await this.subtaskService.update(id,createSubtaskDto,user);
     }
 
     @Delete(':id')
     async delete(@Param('id') id:string){
+        if (!id) {
+            throw new BadRequestException('Subtask ID is required');
+        }
         return await this.subtaskService.deleteSubtaskbyid(id);
     }
